add remaining implementations, notione note generation is still the orchestrator
